+++ pytket/tests/predicates_test.py
@@ -1035,7 +1035,7 @@ def test_greedy_pauli_synth() -> None:
         rega[0], regb[0]
     ).SWAP(regb[1], rega[0])
     d = circ.copy()
-    assert GreedyPauliSimp(0.5, 0.5, thread_timeout=10, trials=5).apply(d)
+    assert GreedyPauliSimp(0.5, 0.5, timeout=10, trials=5).apply(d)
 
     assert np.allclose(circ.get_unitary(), d.get_unitary())
     assert d.name == "test"
@@ -1082,7 +1082,7 @@ def test_greedy_pauli_synth() -> None:
         GreedyPauliSimp().apply(circ)
     err_msg = "Predicate requirements are not satisfied"
     assert err_msg in str(e.value)
-    # large circuit that doesn't complete within thread_timeout argument
+    # large circuit that doesn't complete within timeout argument
     c = Circuit(13)
     for _ in range(20):
         for i in range(13):
@@ -1090,7 +1090,7 @@ def test_greedy_pauli_synth() -> None:
                 c.CX(i, j)
                 c.Rz(0.23, j)
             c.H(i)
-    assert not GreedyPauliSimp(thread_timeout=1).apply(c)
+    assert not GreedyPauliSimp(timeout=1).apply(c)
     assert GreedyPauliSimp().apply(c)