+++ tket/include/tket/Predicates/PassGenerators.hpp
@@ -355,7 +355,7 @@ PassPtr gen_special_UCC_synthesis(
  * @param max_tqe_candidates
  * @param seed
  * @param allow_zzphase
- * @param thread_timeout
+ * @param timeout
  * @param only_reduce
  * @param trials
  * @return PassPtr
@@ -363,9 +363,8 @@ PassPtr gen_special_UCC_synthesis(
 PassPtr gen_greedy_pauli_simp(
     double discount_rate = 0.7, double depth_weight = 0.3,
     unsigned max_lookahead = 500, unsigned max_tqe_candidates = 500,
-    unsigned seed = 0, bool allow_zzphase = false,
-    unsigned thread_timeout = 100, bool only_reduce = false,
-    unsigned trials = 1);
+    unsigned seed = 0, bool allow_zzphase = false, unsigned timeout = 100,
+    bool only_reduce = false, unsigned trials = 1);
 
 /**
  * Generate a pass to simplify the circuit where it acts on known basis states.