+++ pytket/docs/changelog.rst
@@ -12,7 +12,7 @@ Features:
   conditions.
 * Add `custom_deserialisation` argument to `BasePass` and `SequencePass` 
   `from_dict` method to support construction of `CustomPass` from json.
-* Add `thread_timeout`, `only_reduce`, and `trials` arguments 
+* Add `timeout`, `only_reduce`, and `trials` arguments 
   to `GreedyPauliSimp`.
 * Add option to not relabel `ClassicalExpBox` when calling `rename_units`
   and `flatten_registers`